server: handle port 0
Currently, if a user selects port 0 (as advertised in the help menu), there is
no way for the user to trivially find ttyd's port. This commit resolves the
issue by logging the port to the console and by opening a URL with the correct
port information if a user has selected such an option.